% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2poly_a.R
\name{poly2poly_a}
\alias{poly2poly_a}
\title{poly2poly_a -- Polygons to Polygons: area}
\usage{
poly2poly_a(key, value, region, sleep = 1, binary = F, multi = T)
}
\arguments{
\item{key}{character vector of the key values required by OpenStreetMaps. See \link[osmdata]{available_features} for all the possible key values.}

\item{value}{character vector of the tag values matching the key values. See \link[osmdata]{available_tags} for all the possible tag values.}

\item{region}{spatial polygon object. Object of class:  \code{SpatialPolygonsDataFrame}, \code{SpatialPolygons}, \code{SpatialGridDataFrame}, \code{SpatialGrid}, \code{Polygon}, \code{Polygons}, \code{sf}, or \code{sfc}. The polygons, to which the points will be aggregated.}

\item{sleep}{numeric. The waiting time (in seconds) between deriving the next OSM feature. Increasing its value might solve some problems with accessing the OSM server.}

\item{binary}{binary. If \code{TRUE}, the result will be provided in binary form, i.e. the result will be equal to 0, if there are no objects labeled with the provided key and value types, and 1 otherwise. (optional)}

\item{multi}{binary. If \code{TRUE}, the \code{osm_multipolygons} will be decoded into \code{osm_polygons}, \code{osm_multilines} into \code{osm_lines}, and considered along with polygons and linestrings. (optional)}
}
\value{
a data frame object. Each row represents one polygon provided with the region argument. Each column is one aggregated OpenStreetMaps feature.
}
\description{
A function that calculates the total area of \code{osm_polygons} located within provided polygons. Data is derived from OpenStreetMaps database.
}
\author{
Ewa Dobrowolska
}
