% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2pts_a.R
\name{poly2pts_a}
\alias{poly2pts_a}
\title{poly2pts_a -- Polygons to Points: area}
\usage{
poly2pts_a(key, value, pts, radius, sleep = 1, binary = F, multi = T)
}
\arguments{
\item{key}{character vector of the key values required by OpenStreetMaps. See \link[osmdata]{available_features} for all the possible key values.}

\item{value}{character vector of the tag values matching the key values. See \link[osmdata]{available_tags} for all the possible tag values.}

\item{pts}{spatial points object. Object of class: \code{SpatialPointsDataFrame}, \code{SpatialPoints}, \code{SpatialPixels}, \code{SpatialPixelsDataFrame}, \code{sf}, \code{sfc}. The points, around  which the features will be aggregated.}

\item{radius}{numeric. The radius, within which the features will be aggregated.}

\item{sleep}{in seconds. The waiting time between deriving the next OSM feature. Increasing its value might solve some problems with accessing the OSM server.}

\item{binary}{binary. If \code{TRUE}, the result will be provided in binary form, i.e. the result will be equal to 0, if there are no objects labeled with the provided key and value types, and 1 otherwise. (optional)}

\item{multi}{binary. If \code{TRUE}, the \code{osm_multipolygons} will be decoded into \code{osm_polygons}, \code{osm_multilines} into \code{osm_lines}, and considered along with polygons and linestrings. (optional)}
}
\value{
A data frame object. Each row represents one point provided by the pts argument. Each column is one aggregated OpenStreetMaps feature.
}
\description{
A function that calculates the total area of osm_polygons located within the neighbourhood of provided points. Data is derived from OpenStreetMaps database.
}
\author{
Ewa Dobrowolska
}
