% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts2poly.R
\name{pts2poly}
\alias{pts2poly}
\title{pts2poly -- Points to Polygons}
\usage{
pts2poly(key, value, region, sleep = 1, binary = F, outside = F, multi = F)
}
\arguments{
\item{key}{character vector of the key values required by OpenStreetMaps. See \link[osmdata]{available_features} for all the possible key values.}

\item{value}{character vector of the tag values matching the key values. See \link[osmdata]{available_tags} for all the possible tag values.}

\item{region}{spatial polygon object. Object of class:  \code{SpatialPolygonsDataFrame}, \code{SpatialPolygons}, \code{SpatialGridDataFrame}, \code{SpatialGrid}, \code{Polygon}, \code{Polygons}, \code{sf}, or \code{sfc}. The polygons, to which the points will be aggregated.}

\item{sleep}{numeric. The waiting time (in seconds) between deriving the next OSM feature. Increasing its value might solve some problems with accessing the OSM server.}

\item{binary}{binary. If \code{TRUE}, the result will be provided in binary form, i.e. the result will be equal to 0, if there are no objects labeled with the provided key and value types, and 1 otherwise. (optional)}

\item{outside}{binary. If \code{TRUE}, only osm_points located outside osm_polygons will be considered. (optional)}

\item{multi}{binary. If \code{TRUE}, the \code{osm_multipolygons} will be decoded into \code{osm_polygons}, \code{osm_multilines} into \code{osm_lines}, and considered along with polygons and linestrings. (optional)}
}
\value{
A data frame object. Each row represents one polygon provided by the region argument. Each column is one aggregated OpenStreetMaps feature.
}
\description{
A function that calculates the total number of points located within provided polygons. Points are derived from OpenStreetMaps database.
}
\author{
Ewa Dobrowolska
}
